refactor: split vehicle armor tab modules
